backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.2.0
Pillow==10.1.0
qrcode==7.4.2
reportlab==4.0.4
openpyxl==3.1.2
channels==4.0.0
channels-redis==4.1.0

qrcode[pil]
openpyxl
reportlab
Pillow